style: rename `outpututils` package to `outpututil`
